feat(xastro, staking): refactoring.
