+++ contracts/xastro_token/src/lib.rs
@@ -1,2 +1,5 @@
 pub mod contract;
 pub mod state;
+
+#[cfg(test)]
+mod testing;
